Leaderboards
A new problem that kind of works whether candidates want to try and
visualise it or treat it just as a data modelling and service creation
exercise.

It is currently quite short to fit a 45 minute timeslot but it has a
number of ways it can be expanded if needed.